solvers/requirements.txt
Unidecode
BeautifulSoup
nltk.corpus.stopwords
nltk.tag.perceptron.PerceptronTagger